+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Upcoming
+* Added a new `insanelib:feature_enabled` condition for recipes and advancements
 * Features no longer require the 3 parameter constructor. `Feature#init` can be overridden instead if needed
 * Added `Module#getFeature` by name
 * Cleaned up Feature and Module creation